Pi4LCD/requirements.txt
RPLCD
RPi.GPIO
requests
